+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### New features
+
+* Profiler option
+
 ## [v0.3.0](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.3.0) (2017-01-23)
 
 ### New features